Implement Unix control socket
This allows to control the verbosity of the process [none, verbose,
debug] and allows to add/remove interfaces during runtime. A new
dependency to json-c is added as get_meshifs will format the list of
interfaces as json.
