+++ Dockerfile
@@ -7,9 +7,7 @@ RUN DEBIAN_FRONTEND=noninteractive \
     echo 'deb https://dl.ffm.freifunk.net/debian-packages sid main ## FFFFM APT repo' > /etc/apt/sources.list.d/ffffm.list && \
     apt-get update && \
     apt-get -y upgrade && \
-    apt-get -y install \
-        build-essential \
-        git && \
+    apt-get -y install build-essential libjson-c-dev git && \
     apt-get clean autoclean && \
     apt-get autoremove && \
     rm -rf /var/lib/apt/lists/*
